+++ test/mockContracts/VotingToChangeProxyAddressMock.sol
@@ -45,7 +45,7 @@ contract VotingToChangeProxyAddressMock is VotingToChangeMock, VotingToChangePro
         totalVoters = _getTotalVoters(_id);
         progress = _getProgress(_id);
         isFinalized = _getIsFinalized(_id);
-        quorumState = getQuorumState(_id);
+        quorumState = uint8(getQuorumState(_id));
         index = getIndex(_id);
         minThresholdOfVoters = getMinThresholdOfVoters(_id);
         proposedValue = _getProposedValue(_id);
